# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/create": {
            "post": {
                "summary": "Création d'un utilisateur",
                "operationId": "25d0b9760a3cb4c905a7fab64035aa8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "role",
                                    "telephone",
                                    "password",
                                    "name"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "dok@gmail.com"
                                    },
                                    "role": {
                                        "description": "Rôle de l'utilisateur",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "telephone": {
                                        "description": "Numéro de téléphone de l'utilisateur",
                                        "type": "string",
                                        "example": "675656545"
                                    },
                                    "password": {
                                        "description": "Le mot de passe de l'utilisateur",
                                        "type": "string",
                                        "format": "password",
                                        "example": "doknjhyt562@"
                                    },
                                    "name": {
                                        "description": "Nom de l'utilisateur",
                                        "type": "string",
                                        "example": "danie"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur enregistré avec succès"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/details": {
            "get": {
                "summary": "Afficher tous les utilisateurs",
                "operationId": "442b7be86dae1d219431d6ab19fb6e8d",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/search": {
            "get": {
                "summary": "Rechercherche dun utilsateur",
                "operationId": "c1c272da9240904a4d3689138b2933c8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the user to search for",
                        "required": false,
                        "schema": {
                            "type": "email"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/update": {
            "post": {
                "summary": "mise a  d'un utilisateur",
                "operationId": "9c3d017f8fb917368374dba7856868e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "dok@gmail.com"
                                    },
                                    "role": {
                                        "description": "Rôle de l'utilisateur",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "telephone": {
                                        "description": "Numéro de téléphone de l'utilisateur",
                                        "type": "string",
                                        "example": "675656545"
                                    },
                                    "password": {
                                        "description": "Le mot de passe de l'utilisateur",
                                        "type": "string",
                                        "format": "password",
                                        "example": "doknjhyt562@"
                                    },
                                    "name": {
                                        "description": "Nom de l'utilisateur",
                                        "type": "string",
                                        "example": "danie"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur update"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/login": {
            "post": {
                "summary": "Creation d'un login",
                "operationId": "cf7d17e6b88efbe4070506876d0a8483",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'utilisateur",
                                        "type": "email",
                                        "example": "dok@gmail.com"
                                    },
                                    "password": {
                                        "description": "le mot de passe de l'utilisateur",
                                        "type": "paswword",
                                        "example": "doknjhyt562@"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur enregistré avec succès"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}